refactor(silentpayments): extract script template matching to utils/script
