+++ silentpayments/src/utils.rs
@@ -6,6 +6,8 @@
 pub(crate) mod hash;
 #[cfg(feature = "receiving")]
 pub mod receiving;
+#[cfg(any(feature = "sending", feature = "receiving"))]
+pub(crate) mod script;
 #[cfg(feature = "sending")]
 pub mod sending;
